% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyDistOrd.R
\name{MyDistOrd}
\alias{MyDistOrd}
\title{Function to re-order a symmetric additive relationship matrix created by \code{sommer::A.mat} following the algorithm in heatmap().}
\usage{
MyDistOrd(x)
}
\arguments{
\item{x}{Symmetric additive relationship matrix (as created by \code{sommer::A.mat})}
}
\value{
Provides the sorted matrix.
}
\description{
Function to re-order a symmetric additive relationship matrix created by \code{sommer::A.mat} following the algorithm in heatmap().
}
\examples{
require(sommer)
x <- matrix(rnorm(50), 10, 10, 
           dimnames = list(paste("g", 1:10, sep=""), 
           paste("t", 1:10, sep=""))) 
## convert to additive relationship matrix
x <- sommer::A.mat(x)
hv <- heatmap(x)
heatmap.sorted <- x[rev(hv$rowInd), rev(hv$colInd)]
sorted <- MyDistOrd(x)
## to show that the result of heatmap() and this function are the same
heatmap.sorted == sorted
}
